configuracion jwt

+++ src/main/java/com/procesos/negocio/andres/controllers/UsuarioController.java
@@ -3,6 +3,7 @@ package com.procesos.negocio.andres.controllers;
 import com.procesos.negocio.andres.models.Usuario;
 import com.procesos.negocio.andres.repository.UsuarioRepository;
 import com.procesos.negocio.andres.services.UsuarioService;
+import com.procesos.negocio.andres.utils.JWTUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -19,9 +20,15 @@ public class UsuarioController {
 //controlador
     @Autowired
     private UsuarioService usuarioService;
+    @Autowired
+    private JWTUtil jwtUtil;
 
     @GetMapping(value = "/usuario/{id}")
-    public ResponseEntity getUsuario(@PathVariable long id){
+    public ResponseEntity getUsuario(@PathVariable long id,@RequestHeader(value = "Authorization")String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("token no valido");
+
+        }
         return usuarioService.getUserById(id);
 
     }
@@ -31,36 +38,61 @@ public class UsuarioController {
 
     }
     @GetMapping("/usuarios")
-    public ResponseEntity listaUsuarios(){
+    public ResponseEntity listaUsuarios(@RequestHeader(value = "Authorization")String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("token no valido");
+
+        }
         return usuarioService.allUsers();
 
     }
 
     @GetMapping("/usuario/{nombre}/{apellidos}")
-    public ResponseEntity listarPorNombreApellidos(@PathVariable String nombre, @PathVariable String apellidos){
+    public ResponseEntity listarPorNombreApellidos(@PathVariable String nombre
+            , @PathVariable String apellidos,@RequestHeader(value = "Authorization")String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("token no valido");
+
+        }
         return usuarioService.allUsersByNameAndLastName(nombre,apellidos);
     }
 
     @GetMapping("/usuario/apellidos/{apellidos}")
-    public ResponseEntity listarPorApellidos( @PathVariable String apellidos){
+    public ResponseEntity listarPorApellidos( @PathVariable String apellidos,@RequestHeader(value = "Authorization")String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("token no valido");
+
+        }
         return usuarioService.allUsersByLastName(apellidos);
 
 
 
     }
     @GetMapping("/usuario/nombre/{nombre}")
-    public ResponseEntity listarPorNombre( @PathVariable String nombre){
+    public ResponseEntity listarPorNombre( @PathVariable String nombre,@RequestHeader(value = "Authorization")String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("token no valido");
+
+        }
         return usuarioService.allUsersByName(nombre);
 
     }
 
     @PutMapping("/usuario/{id}")
-    public ResponseEntity editarUsuario(@PathVariable Long id ,@Valid @RequestBody  Usuario usuario){
+    public ResponseEntity editarUsuario(@PathVariable Long id ,@Valid @RequestBody  Usuario usuario,@RequestHeader(value = "Authorization")String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("token no valido");
+
+        }
         return usuarioService.editUser(id,usuario);
     }
 
     @DeleteMapping("/usuario/{id}")
-    public ResponseEntity eliminarUsuario(@PathVariable Long id){
+    public ResponseEntity eliminarUsuario(@PathVariable Long id,@RequestHeader(value = "Authorization")String token){
+        if(jwtUtil.getKey(token) == null){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("token no valido");
+
+        }
         return usuarioService.deleteUser(id);
 
     }
